refactor: Modularize runtime into directory structure (Phase 1)
Break down the monolithic runtime.rs (~4100 lines) into organized modules
for improved maintainability and separation of concerns:

- runtime/mod.rs: Core StreamRuntime struct, add_processor, spawn_processor_thread
- runtime/state.rs: RuntimeState, ProcessorStatus, WakeupEvent enums
- runtime/types.rs: Connection, RuntimeProcessorHandle, type aliases
- runtime/lifecycle.rs: start, stop, pause, resume, restart methods
- runtime/connections.rs: connect, disconnect, wire_pending_connections

Key changes:
- Each module has single responsibility and clear boundaries
- All 218 lib tests pass
- Examples compile successfully (camera-display, microphone-reverb-speaker)
- Added missing remove_processor method for API completeness
- Updated integration tests to use new processor handle API
- Fixed ConnectionId re-export conflict between bus and runtime modules

Note: Python feature has pre-existing build errors (unrelated to this refactor)
